refactor: move processor and pipeline from service to domain
- Processor and Pipeline is domain layer.
