+++ source-connector/src/main/java/sourceconnector/domain/processor/AbstractMapperProcessor.java
@@ -1,4 +1,4 @@
-package sourceconnector.service.processor;
+package sourceconnector.domain.processor;
 
 public abstract class AbstractMapperProcessor<Log> extends BaseProcessor<Log> {
