+++ source-connector/src/main/java/sourceconnector/domain/processor/Processor.java
@@ -1,4 +1,4 @@
-package sourceconnector.service.processor;
+package sourceconnector.domain.processor;
 
 public interface Processor<I, R> {
   R process(I record);
