+++ source-connector/src/main/java/sourceconnector/domain/processor/impl/EmptyFilterProcessor.java
@@ -1,7 +1,7 @@
-package sourceconnector.service.processor.impl;
+package sourceconnector.domain.processor.impl;
 
 import sourceconnector.domain.log.Log;
-import sourceconnector.service.processor.AbstractFilterProcessor;
+import sourceconnector.domain.processor.AbstractFilterProcessor;
 
 /**
  * Filter out the empty Log
